Abstract QArray into base class and MaterializedQArray subclass
