+++ dynamiqs/qarrays/__init__.py
@@ -1,3 +1,4 @@
+from .materialized_qarray import *  # registers MaterializedQArray (not exported)
 from .qarray import *
 from .qarray import QArrayLike  # excluded from documentation
 from .utils import *
